Extract the MCP session registry into sessionRegistry.ts
Move SessionEntry/SessionRegistry, InFlightCounter/createInFlightCounter, and
markSessionActive/markSessionIdle out of the 1300-line streamableHttp.ts into a
self-contained module. Pure code move — no behaviour change; the four test suites
that used these symbols import them from the new module.
